First version of redirect counter
Need to start the tweet stream manually. Need to figure out how to get
it going and fire off events for each tweet so the gen server isn't
blocked and we can stop it if we want.

Also need to figure out how to prevent crashes from the url counter
worker processes that aren't interesting from being output to the
terminal